Move score calculation into the FinalScore struct
